Start work on Ticket #55, implementing a set count of player lives.
This is far from done, right now it only tells you how many lives you
have left and then keeps going negative as there is nothing setup to
teleport you elsewhere when you run out of lives. That is going to be
a whole new set of code which will be done later.